+++ plugin/src/main/java/jp/gr/java_conf/spica/plugin/gradle/jacoco/internal/domain/coverage/model/ClassNameFilterString.java
@@ -1,0 +1,21 @@
+package jp.gr.java_conf.spica.plugin.gradle.jacoco.internal.domain.coverage.model;
+
+public class ClassNameFilterString {
+
+  private final String filterString;
+
+  public ClassNameFilterString(String filterString) {
+    this.filterString = filterString;
+  }
+
+  public boolean isRegex() {
+    if (filterString.isEmpty()) {
+      return false;
+    }
+    return filterString.charAt(0) == '^' && filterString.charAt(filterString.length() - 1) == '$';
+  }
+
+  public String filterString() {
+    return filterString;
+  }
+}